refactor: move modalReducer into ModalLayer, separate out modal window
